fixing switch for community projects / posts
