+++ frontend/src/App.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
+import { BrowserRouter, Route } from 'react-router-dom';
 import Header from './Components/Header/Header';
 import './App.css';
 import { fetchUser } from './redux/auth';
@@ -12,10 +12,9 @@ import Login from './Components/Authentication/Login';
 import CreateProject from './Components/Projects/CreateProject';
 import Profile from './Components/UserProfile/Profile';
 
-import CommunityProjects from './Component/CommunityProjects/CommunityProjects';
+import CommunityProjects from './Components/CommunityProjects/CommunityProjects';
 import Dashboard from './Components/Dashboard/Dashboard';
 import CommunityPosts from './Components/CommunityPosts/CommunityPosts';
-
 
 class App extends Component {
   constructor() {
@@ -42,26 +41,24 @@ class App extends Component {
       <div>
         <BrowserRouter>
           <div className="App">
-
-            <Header 
+            <Header
               displayPosts={this.state.displayPosts}
               onClickPosts={this.onClickPosts}
             />
-            <Dashboard />
-            <Switch>
+            <Dashboard
+              displayPosts={this.state.displayPosts}
+              onClickPosts={this.onClickPosts}
+            />
             <Route exact path="/register" component={Register} />
             <Route exact path="/login" component={Login} />
             <Route exact path="/create-project" component={CreateProject} />
             <Route exact path="/profile" component={Profile} />
-            <Route exact path="/communityprojects" component={CommunityProjects}
-            </Switch>
             <Route
               exact
               path="/community-projects"
               component={CommunityProjects}
             />
             <Route exact path="/community-posts" component={CommunityPosts} />
-
           </div>
         </BrowserRouter>
       </div>
